【UX增强】修复HyperLink中setColor的默认值不正确的问题 Signed-off-by: Xing Guang <xingguang1@huawei.com>

+++ frameworks/bridge/declarative_frontend/engine/jsi/components/arkts_native_common_modifier.cpp
@@ -2137,7 +2137,7 @@ void SetAllowDrop(NodeHandle node, const char** allowDropCharArray, int32_t leng
     std::set<std::string> allowDropSet;
     allowDropSet.clear();
     std::string allowDropStr;
-    for (size_t i = 0; i < length; i++) {
+    for (int32_t i = 0; i < length; i++) {
         allowDropStr = allowDropCharArray[i];
         allowDropSet.insert(allowDropStr);
     }
